pyproject.toml
[project]
name = "hollow-knight-save"
version = "0.1.0"
description = "A tool for reading and modifying Hollow Knight (and Silksong) save files."
readme = "README.md"
authors = [
    { name = "hotwords123", email = "hotwords123@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pycryptodome>=3.23.0",
]

[project.scripts]
hollow-knight-save = "hollow_knight_save.main:run"

[build-system]
requires = ["uv_build>=0.8.18,<0.9.0"]
build-backend = "uv_build"
